+++ src/ray/observability/tests/BUILD.bazel
@@ -6,7 +6,7 @@ ray_cc_test(
     srcs = ["open_telemetry_metric_recorder_test.cc"],
     tags = ["team:core"],
     deps = [
-        "//src/ray/telemetry:open_telemetry_metric_recorder",
+        "//src/ray/observability:open_telemetry_metric_recorder",
         "@com_google_googletest//:gtest_main",
     ],
 )
