+++ src/ray/stats/BUILD.bazel
@@ -14,7 +14,7 @@ ray_cc_library(
     ],
     deps = [
         "//src/ray/common:ray_config",
-        "//src/ray/telemetry:open_telemetry_metric_recorder",
+        "//src/ray/observability:open_telemetry_metric_recorder",
         "//src/ray/util",
         "//src/ray/util:logging",
         "//src/ray/util:size_literals",
